+++ unpythonic/tests/test_dispatch.py
@@ -5,7 +5,7 @@ from ..test.fixtures import session, testset, returns_normally
 
 import typing
 from ..fun import curry
-from ..dispatch import generic, generic_for, typed
+from ..dispatch import generic, generic_addmethod, typed
 
 @generic
 def zorblify(x: int, y: int):
@@ -91,11 +91,11 @@ def runtests():
         test[gargle(42, 6.022e23, "hello") == "int, float, str"]
         test[gargle(1, 2, 3) == "int"]  # as many as in the [int, float, str] case
 
-    with testset("@generic_for"):
+    with testset("@generic_addmethod"):
         @generic
         def f1(x: typing.Any):
             return False
-        @generic_for(f1)
+        @generic_addmethod(f1)
         def f2(x: int):
             return x
         test[f1("hello") is False]
@@ -103,8 +103,8 @@ def runtests():
 
         def f3(x: typing.Any):  # not @generic!
             return False
-        with test_raises[TypeError, "should not be able to @generic_for a non-generic function"]:
-            @generic_for(f3)
+        with test_raises[TypeError, "should not be able to @generic_addmethod a non-generic function"]:
+            @generic_addmethod(f3)
             def f4(x: int):
                 return x
 
@@ -269,12 +269,12 @@ def runtests():
         # Since these are in the same lexical scope as the original definition of the
         # generic function `flippable`, we could do this using `@generic`, but
         # later extensions (which are the whole point of traits) will need to specify
-        # on which function the new methods are to be registered, using `@generic_for`.
+        # on which function the new methods are to be registered, using `@generic_addmethod`.
         # So let's do that to show how it's done.
-        @generic_for(flippable)
+        @generic_addmethod(flippable)
         def flippable(x: str):  # noqa: F811
             return IsFlippable()
-        @generic_for(flippable)
+        @generic_addmethod(flippable)
         def flippable(x: int):  # noqa: F811
             return IsNotFlippable()
 
@@ -289,7 +289,7 @@ def runtests():
         def flip(x: typing.Any):
             return flip(flippable(x), x)
 
-        # Implementation of `flip`. Same comment about `@generic_for` as above.
+        # Implementation of `flip`. Same comment about `@generic_addmethod` as above.
         #
         # Here we provide one implementation for "flippable" objects and another one
         # for "nonflippable" objects. Note this dispatches regardless of the actual
@@ -299,10 +299,10 @@ def runtests():
         # We could also add methods for specific types if needed. Note this is not
         # Julia, so the first matching definition wins, instead of the most specific
         # one.
-        @generic_for(flip)
+        @generic_addmethod(flip)
         def flip(traitvalue: IsFlippable, x: typing.Any):  # noqa: F811
             return x[::-1]
-        @generic_for(flip)
+        @generic_addmethod(flip)
         def flip(traitvalue: IsNotFlippable, x: typing.Any):  # noqa: F811
             raise TypeError(f"{repr(x)} is IsNotFlippable")